merge: 기능 통합 및 정리
