+++ src/main/java/com/tag/prietag/model/Template.java
@@ -26,6 +26,12 @@ public class Template extends TimeStamped {
     private boolean isDeleted;
 
     @Builder
+    public Template(Long id, User user, String mainTitle) {
+        this.id = id;
+        this.user = user;
+        this.mainTitle = mainTitle;
+        this.isDeleted = false;
+    }
     public Template(User user, String mainTitle) {
         this.user = user;
         this.mainTitle = mainTitle;
