+++ src/main/resources/application-dev.yml
@@ -35,7 +35,7 @@ logging:
     '[org.hibernate.type]': TRACE # 콘솔 쿼리에 ? 에 주입된 값 보기
 
 jwt:
-  secret: fdsafdaffds
+  secret: adfsafddsf
 
 # S3
 cloud:
