+++ src/main/resources/application-prod.yml
@@ -21,6 +21,9 @@ logging:
   level:
     '[com.tag.prietag]': INFO # DEBUG 레벨부터 에러 확인할 수 있게 설정하기
 
+jwt:
+  secret: ${SECRET}
+
 # S3
 cloud:
   aws:
