+++ src/main/resources/application.yml
@@ -3,6 +3,3 @@ spring:
     active:
       - dev
     include: oauth
-
-jwt:
-  secret: ${SECRET}
