+++ src/test/java/com/tag/prietag/repository/CustomerLogRepositoryTest.java
@@ -46,7 +46,7 @@ public class CustomerLogRepositoryTest {
         User user = User.builder()
                 .email("dbs@naver.com")
                 .username("Lee")
-                .role(User.Role.USER)
+                .role(User.RoleEnum.USER)
                 .publishId(6L)
                 .build();
         userRepository.save(user);
