+++ src/test/java/com/tag/prietag/repository/PublishLogRepositoryTest.java
@@ -41,7 +41,7 @@ public class PublishLogRepositoryTest {
         User user = User.builder()
                 .email("dbs@naver.com")
                 .username("Lee")
-                .role(User.Role.USER)
+                .role(User.RoleEnum.USER)
                 .publishId(6L)
                 .build();
         userRepository.save(user);
