+++ src/test/java/com/tag/prietag/service/TemplateServiceTest.java
@@ -73,7 +73,7 @@ public class TemplateServiceTest {
                 .email("dasfd@naver.com")
                 .username("Lee")
                 .publishId(1L)
-                .role(User.Role.USER)
+                .role(User.RoleEnum.USER)
                 .build();
         template = Template.builder()
                 .id(1L)
